fix: rename slint_maplibre to slint_maplibre_headless files
